+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Unreleased
+
+* Add `Store#disconnect`
+
 ### 0.2.0 (2016-11-14)
 
 * Add `Store#padded_shard` to get the formatted shard number for a shardable value